requirements.txt
# Dependencies
fastapi
uvicorn
google-genai
python-dotenv
sqlalchemy
pydantic
